Stop using bootloader-provided OS version & patchlelvel
Keymaster2 required that the bootloader provide OS version and
patchlevel retrieved from boot.img, and that the trusted app compare
those with values provided by the system, from the read-only system
properties, and refuse to function if the boot and system values
differ.  This requirement effectively forced boot and system images to
always be updated together.

This made sense when it was defined in the pre-Treble world, but no
longer does, since the small security benefit it provides is more than
outweighed by the function and security cost of blocking system-only
updates and upgrades.  So the decision has been made to relax this
requirement.

This CL changes the Trusty Keymaster implementation so it ignores the
OS version and patchlevel information provided by the bootloader, and
uses the information provided by the system, the first time the HAL is
loaded.  This means that keys will only be bound to system-reported
values.

Bug: 79725215
Change-Id: Ie286e6e9e6118e4118a97fdfc1c4ccb4ffb38483